[MODIFY] users 파일들에서 cart 분리
