+++ services/cartService.js
@@ -1,0 +1,9 @@
+const cartDao = require("../models/cartDao");
+
+const cartInfo = async (userId) => {
+  return await cartDao.getCart(userId);
+};
+
+module.exports = {
+  cartInfo,
+};
